_co_send(): do not rely on select to determine if we can send data. Since the socket is in non-blocking mode, just try to send and catch the necessary exceptions.
This fixes a conjectured problem where select reports that the socket is
writable, but it actually isn't (because select isn't 100% reliable, and
because in case of an ssl socket we don't write to the fd directly). In
this case we would have gotten a crash before, because we didn't catch
the necessary exceptions.