_try_fill_buffer(): fix handling of empty and full buffer.
As written, the if clause was always false, so if a buffer was
already filled to capacity, we got an AssertionError. For an
empty buffer, we did not reset the rbuf.b counter which reduced
performance.

--HG--
extra : histedit_source : aad5f560f4ad7254e4cbe2a9445a91fea43c483a