_co_read_chunked(): removed TODO
We could set _in_remaining to a value bigger than the remainder of the
current chunk. However, in that case we would need to copy the extra
data from the provided buffer into the connection buffer. It is not clear
if that is faster than calling sock.recv_into a second time, but it
would make the code significantly more complex.