Use select instead of poll when poll is unavailable (e.g., on Win32).
Continue to use poll when it is available, to avoid "filedescriptor out
of range in select()" errors. (Win32's select takes a list of handles,
not a bit vector, so this error can't occur.)

With this change, dugong appears to work fine on Windows in my limited
testing. Without this change, it crashes on the import of POLLIN and
POLLOUT.

Clients of dugong that import and use the POLLIN and POLLOUT constants
explicitly will still fail on Windows. The fix is to import them from
dugong instead of select. It would be better to use other constants
(such as 'r' and 'w') but that would break API compatibility.